Call withdraw on ForeignGivethBridge in _collect()
This eliminates the need for the user to have to call withdraw on the
bridge, simplifying the steps required to get paid.

Removed ether support because this contract is ment to be used in
conjuntion with the ForeignGivethBridge, which will proved a wrapped eth
token for the home network eth.